References

OrthographicProjection.ParametersExtraction — Type

ParametersExtraction

Description:

outputfile: Output filename for point cloud

potreedirs: Potree projects

model: Cuboidal LAR model

mainHeader: header of point cloud file las

Fields

outputfile::String
potreedirs::Array{String,1}
model::Lar.LAR
mainHeader::LasIO.LasHeader

OrthographicProjection.ParametersOrthophoto — Type

ParametersOrthophoto

Description:

PO: Projection plane and sight direction. Option: "XY+","XY-","XZ+","XZ-","YZ+","YZ-"

outputimage: Output filename for image

outputfile: Output filename for point cloud

potreedirs: Potree projects

model: Cuboidal LAR model

coordsystemmatrix: Rotation matrix related to PO

RGBtensor: Image tensor

rasterquote: Matrix of heigth (z-buffer)

GSD: Ground sampling distance

refX: x coordinate of landmark

refY: y coordinate of landmark

pc: if true extract point cloud

ucs: user coordinate system

mainHeader: header of point cloud file las

Fields

PO::String
outputimage::String
outputfile::String
potreedirs::Array{String,1}
model::Lar.LAR
coordsystemmatrix::Array{Float64,2}
RGBtensor::Array{Float64,3}
rasterquote::Array{Float64,2}
GSD::Float64
refX::Float64
refY::Float64
pc::Bool
ucs::Matrix
mainHeader::LasIO.LasHeader

OrthographicProjection.PO2matrix — Function

PO2matrix(PO::String, UCS=Matrix{Float64}(Lar.I,4,4)::Matrix)

Observation point. Valid input:

"XY+": Top view

"XY-": Bottom view

"XZ+": Back view

"XZ-": Front view

"YZ+": Left view

"YZ-": Right view

OrthographicProjection.add_point — Method

add_point!(params::Union{ParametersExtraction,ParametersOrthophoto}, laspoint::LasIO.LasPoint, h::LasIO.LasHeader, s::IOStream, n::Int64)

Add new point to segmented point cloud.

OrthographicProjection.dfs — Method

dfs(t::DataStructures.Trie{String},

params::Union{ParametersOrthophoto,ParametersExtraction}, s::Union{Nothing,IOStream},n::Int64,nfiles::Int64,l::Int64)

Depth search first.

OrthographicProjection.extract_models — Method

get_sections(
	txtpotreedirs::String,
	project_name::String,
	proj_folder::String,
	bbin::Union{AABB,String},
	models::Array{Lar.LAR,1})

For each model in models extracts and saves the clipped point cloud.

OrthographicProjection.get_parallel_sections — Method

getparallelsections( txtpotreedirs::String, projectname::String, projfolder::String, bbin::Union{AABB,String}, step::Float64, plane::Plane, model::Lar.LAR)

Return parallel sections of point cloud.

Input:

A text file containing a list of files to segment

Project name

A folder for saving results

A box model: region of interest

Distance between slices

Plane description of first slice

LAR model of thickness plane

Output:

A file .las for each slice

OrthographicProjection.get_potree_dirs — Method

get_potree_dirs(txtpotreedirs::String)

Return collection of potree directories.

OrthographicProjection.get_quotas — Method

Return quotas and indices of each LAR model thickness plane.

OrthographicProjection.init — Method

Initialize parameters for orthographic projection algorithm.

OrthographicProjection.init — Method

Initialize parameters for segment algorithm.

OrthographicProjection.init_raster_array — Method

init_raster_array(matrix::Array{Float64,2}, GSD::Float64, model::Lar.LAR, BGcolor::Array{Float64,1})

Orthoprojection of model on plane defined by matrix, create a raster image with background color BGcolor.

OrthographicProjection.orthophoto — Method

orthophoto( txtpotreedirs::String, outputimage::String, bbin::Union{String,AABB}, GSD::Float64, PO::String, quota::Union{Float64,Nothing}, thickness::Union{Float64,Nothing}, ucs::Union{String,Matrix{Float64}}, BGcolor::Array{Float64,1}, pc::Bool )

Orthographic projection of point cloud.

Description:

txtpotreedirs: path to Potree project or a text file containing a list of Potree projects

outputimage: output filename for image

bbin: region of interest, aabb ("xmin ymin zmin xmax ymax zmax") or volume (path to json file of described volume)

GSD: ground sampling distance

PO: projection plane and sight direction. Option: "XY+","XY-","XZ+","XZ-","YZ+","YZ-"

ucs: user coordinate system

BGcolor: background color of image

pc: if true saves the segmented point cloud in region of interest

It is possible to limit the region of interest for a section of point cloud. The slice is described with:

quota: distance of plane to origin

thickness: thickness of plane

Put both of them to nothing if you are not interested in a slice.

OrthographicProjection.orthophoto_core — Method

Core

OrthographicProjection.preprocess — Method

preprocess( projectname::String, outputfolder::String, bbin::Union{AABB,String}, p1::Array{Float64,1}, p2::Array{Float64,1}, axis_y::Array{Float64,1}, thickness::Float64 )

Create project folder and create LAR model of thickness plane described by two points and a vector in plane.

OrthographicProjection.preprocess — Method

preprocess( projectname::String, outputfolder::String, bbin::Union{AABB,String}, plane::Plane, thickness::Float64 )

Create project folder and create LAR model of thickness plane described in Hessian normal form.

OrthographicProjection.preprocess — Method

preprocess( projectname::String, outputfolder::String, bbin::Union{AABB,String}, filelistpoints::String, axisy::Array{Float64,1}, thickness::Float64 )

Create project folder and create LAR models of thickness planes described by two points (read in file) and a vector in plane.

OrthographicProjection.saveimage — Method

Save orthoprojection image.

OrthographicProjection.savepointcloud — Method

Save point cloud extracted file .las.

OrthographicProjection.segment — Method

segment(txtpotreedirs::String, output::String, model::Lar.LAR)

Input:

A text file containing a list of files to segment

output LAS filename

model of extraction

Output:

Point cloud LAS

OrthographicProjection.trie_traversal — Method

trie_traversal(params::Union{ParametersExtraction,ParametersOrthophoto},s::Union{Nothing,IOStream})

Trie traversal.

OrthographicProjection.update! — Method

update!(params::Union{ParametersOrthophoto,ParametersExtraction}, file::String, s::Union{Nothing,IOStream}, n::Int64)

Process points in file without further checks.

OrthographicProjection.update_core — Method

update_core(params::ParametersExtraction, laspoint::LasIO.LasPoint, h::LasIO.LasHeader, s, n::Int64)

OrthographicProjection.update_core — Method

OrthographicProjection.updateif! — Method

updateif!(params::Union{ParametersOrthophoto,ParametersExtraction}, file::String, s, n::Int64)

Process all points, in file, falling in region of interest.

OrthographicProjection.trie_traversal — Function

trie_traversal(params::Union{ParametersExtraction,ParametersOrthophoto},s::Union{Nothing,IOStream})

Trie traversal.

OrthographicProjection.dfs — Function

dfs(t::DataStructures.Trie{String},

params::Union{ParametersOrthophoto,ParametersExtraction}, s::Union{Nothing,IOStream},n::Int64,nfiles::Int64,l::Int64)

Depth search first.