References

OrthographicProjection.ParametersExtraction — Type

ParametersExtraction

Constructors

ParametersExtraction(txtpotreedirs::String,
	outputfile::String,
	model::Lar.LAR;
	epsg = nothing::Union{Nothing,Integer}
	)

Input description:

txtpotreedirs: path to Potree project or a text file containing a list of Potree projects

outputfile: output LAS filename

model: cuboidal LAR model

epsg: projection code

Fields

outputfile::String
potreedirs::Array{String,1}
model::Lar.LAR
mainHeader::LasIO.LasHeader

Fields description:

outputfile: Output filename for point cloud

potreedirs: Potree projects

model: Cuboidal LAR model

mainHeader: header of point cloud LAS file

OrthographicProjection.ParametersOrthophoto — Type

ParametersOrthophoto

Constructors

ParametersOrthophoto(
	txtpotreedirs::String,
	outputimage::String,
	bbin::Union{String,AABB},
	GSD::Float64,
	PO::String,
	quota::Union{Float64,Nothing},
	thickness::Union{Float64,Nothing},
	ucs::Union{String,Matrix{Float64}},
	BGcolor::Array{Float64,1},
	pc::Bool
	)

Input description:

txtpotreedirs: path to Potree project or a text file containing a list of Potree projects

outputimage: Output filename for image

bbin: Path to json file or AABB

GSD: Ground sampling distance

PO: Projection plane and sight direction. Option: "XY+","XY-","XZ+","XZ-","YZ+","YZ-"

quote: Distance of plane from the origin

thickness: Thickness of plane

ucs: User coordinates system

BGcolor: Background color image

pc: If true return LAS file of point cloud

Fields

PO::String
outputimage::String
outputfile::String
potreedirs::Array{String,1}
model::Lar.LAR
coordsystemmatrix::Array{Float64,2}
RGBtensor::Array{Float64,3}
rasterquote::Array{Float64,2}
GSD::Float64
refX::Float64
refY::Float64
pc::Bool
ucs::Matrix
header_bb::AABB
mainHeader::LasIO.LasHeader

Fields description:

PO: Projection plane and sight direction. Option: "XY+","XY-","XZ+","XZ-","YZ+","YZ-"

outputimage: Output filename for image

outputfile: Output filename for point cloud

potreedirs: Potree projects

model: Cuboidal LAR model

coordsystemmatrix: Rotation matrix related to PO

RGBtensor: Image tensor

rasterquote: Matrix of heigth (z-buffer)

GSD: Ground sampling distance

refX: x coordinate of landmark

refY: y coordinate of landmark

pc: if true extract point cloud

ucs: user coordinate system

header_bb: segmented point cloud AABB

mainHeader: header of point cloud file las

OrthographicProjection.trie_traversal — Function

trie_traversal(params::Union{ParametersExtraction,ParametersOrthophoto},s::Union{Nothing,IOStream})

Trie traversal. If entire point cloud falls in volume process all files of Potree project else travers trie, depth search first, and process nodes falling in region of interest.

Input:

params: initial parameters

s: nothing or stream file where to save las points

Output:

n: number of processed points

OrthographicProjection.dfs — Function

dfs(t::DataStructures.Trie{String},
params::Union{ParametersOrthophoto,ParametersExtraction},
s::Union{Nothing,IOStream},n::Int64,nfiles::Int64,l::Int64)

Depth search first.